Write test prepared
