+++ server/src/main/java/com/matterhub/server/entities/metrics/DCMDMetric.java
@@ -23,23 +23,6 @@ public abstract class DCMDMetric
 }
 
 
-class DCMDWriteMetric
-        extends DCMDMetric {
-    private final int attributeId;
-
-    public DCMDWriteMetric(short endpointId, int clusterId, int attributeId, long timestamp, DataType dataType,
-                           JsonNode value) {
-        super(endpointId, clusterId, timestamp, dataType, value);
-        this.attributeId = attributeId;
-    }
-
-
-    @Override
-    protected String lastPartOfName() {
-        return "write/" + attributeId;
-    }
-}
-
 class DCMDSubscribeMetric
         extends DCMDMetric {
     private final int attributeId;
